Removing not needed settings field, more translatable strings

+++ inc/utilities.php
@@ -4,7 +4,7 @@
  * @return bool
  */
 function is_generic_member_user() {
-	return get_current_user_id() == get_option( 'ssc_member_generic_user', 0 );
+	return get_current_user_id() == (int) get_option( 'ssc_member_generic_user', 0 );
 }
 
 /**
